Refactor the function adapters to return dicts
This commit refactors the functional command adapters in PR #3 so the
convenience functions return dicts instead of lists. The dicts returned
by execute_function are keyed by the account. The dicts returned by the
regional_execute_function are keyed by the account and region.

In addition, this commit adds documentation to the command adapters and
the convenience functions.