Add new `console` command
This commit adds a new command called `console` to the list of included
commands in awsrun. It generates a temporary sign-in URL for the AWS
Console for the accounts specified. In an enterprise environment where
SAML-based federation is typically used, this allows one to obtain a URL
using the same credentials the user has configured for awsrun.

For example, this is convenient when working in an environment where one
has multiple usernames and it's inconvenient to switch usernames in your
browser for SSO to the console because it is so easy to switch usernames
via the awsrun CLI.

See docstring of command for usage.

Signed-off-by: Kazmier, Peter <peter.kazmier@fmr.com>