Add new `last` command
This commit adds a new command called `last` to the included commands in
awsrun. It is named after the same command in unix. The command pulls
the last N hours of events from CloudTrail for a given account. Specific
start and end times can be provided instead. One can also filter events
using standard AWS CloudTrail filters.

Results can be viewed in a non-interactive manner where they are simply
printed to the console color-coded where events from the same user are
shown in the same color. For more fun though, events can be reviewed in
an interactive TUI (terminal user interface), which also provides access
to the raw JSON for each event.

Signed-off-by: Kazmier, Peter <peter.kazmier@fmr.com>