Include DXGW Id in verbose output of dx_status cmd
The verbose option for the dx_status command provides a list of all VIFs
associated with a Direct Connect connection as well as the VGW Id the VIF
is connected to. VIFs, however, may be connected to DXGWs as well, in
which case, this change will now include the DXGW Id.

Signed-off-by: Kazmier, Peter <peter.kazmier@fmr.com>