Move AWS-specific session providers to a submodule
In preparation for the addition of Azure session providers, it makes
sense to move the AWS implementation to a submodule called `aws`. The
new Azure providers will be in a submodule called `azure`.

We also change the way the default session provider is selected. Before,
it would search for a module called `Profile`, which made sense when
supporting only AWS. We now search for a module called `Default` in the
respective CSP submodule.

Signed-off-by: Kazmier, Peter <peter.kazmier@fmr.com>