Add new AzureCLI account loader plug-in
Because the Azure CLI maintains a local list of all the subscriptions a
user has access to, we can take advantage of that with a custom account
loader that uses the output of `az account list --all` to generate the
possible list of accounts that can be acted upon. In addition, we attach
metadata to each account including optional custom parsing of the
subscription name for even more attributes that can be used with the
metadata account filters (--include and --exclude).

Signed-off-by: Kazmier, Peter <peter.kazmier@fmr.com>