fix: allow reuse of argparse option flags
This change removes the safety valve to prevent command authors from
inadvertently defining command line arguments for their commands that
conflict with the main awsrun CLI flags (such as --account, --include,
etc ...).

If a command author does define the same flag, it will never be reached
as the main argument parser for awsrun will have consumed it first, so
by the time the command's argument parser is run, the argument is gone.

This safety valve was a hack as it overrode the internals of the
ArgumentParser to prevent global reuse of options across any instance of
ArgumentParser.

While the intention was good, the consequences prevented command authors
from reusing option flags across CLI subcommands as each subcommond is
just another ArgumentParser instance.  At the end of the day, a command
author will test their code and realize if they've accidentally used an
option flag defined by awsrun proper.