build!: slim default install with optional deps
Previously, all dependencies needed for awsrun, azurerun, and all of the
bundled sample commands were installed when doing:

    $ pip install awsrun

With this commit, only the bare minimum is installed. Users MUST now
opt-in to the dependencies they might need.  This is a breaking change
as the previous command will no longer suffice.

At a minimum, all users will need to specify either "aws" and/or
"azure". For example, to install awsrun:

    $ pip install "awsrun[aws]"

Or to install dependencies needed for both awsrun and azurerun:

    $ pip install "awsrun[aws,azure]"

Additionally, several of the included sample commands have additional
dependencies that are no longer installed. If you attempt to use one of
those commands, but have not installed the optional deps, you will be
asked to install them with instructions on how to do so. For example, to
install the dependencies for the `last` command (CloudTrail event
viewer):

    $ pip install "awsrun[last]"

To install all of the optional dependencies for the included commands,
you can also do the following (assuming you already installed either
"aws" and/or "azure"):

    $ pip install "awsrun[cmds]"

Signed-off-by: Kazmier, Peter <peter.kazmier@fmr.com>