fix: delay prompting user in URLLoader plug-in
At the time the URLLoader plug-in is instantiated, we don't know if
the URLAccountLoader will even need to make an HTTP call as it may
have cached data. So, if the user has not specified a password via
CLI args, env vars, or config file, we don't want to prematurely
prompt them unless URLAccountLoader really needs to make an HTTP
request.