refactor: move context parameter to run()
Initially, when discussing where to inject the context parameter, it was
added to the consructor of the AccountRunner. However, the more I think
about it, the idea of a runtime context makes more sense to be added to
the AccountRunner.run method, so a different context can be provided for
multiple run() invocations. If we didn't do this, then one would need to
create a new instance of AccountRunner if one wanted to pass a different
context upon each run().

Also, renamed CLIContext back to Context to avoid the repetition of
"cli" when importing given this class exists in the cli module. I.e.
insead of:

    from awsrun.cli import CLIContext

The import is now:

    from awsrun.cli import Context