feat(dx_maint): sort output by start time of event (#38)
Previously the output was sorted by the date the event was published
because AWS does not always provide a start time in events (I have no
idea why). But for a series of maintenance events, sorting by the start
time of the event makes more sense. In the event that AWS does not
provide a start time, then we fallback to the time the event was
published.